trustmile/requirements.txt
aftership==0.1
alembic==0.8.0
amqp==1.4.7
aniso8601==1.0.0
ansible==2.0.2.0
anyjson==0.3.3
apiclient==1.0.2
appnope==0.1.0
awscli==1.7.45
backports.ssl-match-hostname==3.4.0.2
bcrypt==2.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.21
blinker==1.3
boto==2.38.0
boto3==1.3.0
botocore==1.4.8
cairocffi==0.7.2
CairoSVG==1.0.20
cchardet==1.0.0
celery==3.1.19
certifi==2015.9.6.2
cffi==1.5.0
chardet==1.0.1
click==6.6
colorama==0.3.3
cryptography==1.2.1
cssmin==0.1.4
Cython==0.23.4
decorator==4.0.2
dnspython==1.12.0
dnsq==1.1.6
docopt==0.4.0
docutils==0.12
dogpile.cache==0.5.6
dogpile.core==0.4.1
dpath==1.4.0
ecdsa==0.13
enum34==1.1.2
expiringdict==1.1.3
facebook-sdk==0.3.1
flanker==0.4.36
Flask==0.10.1
Flask-Assets==0.7
Flask-Environments==0.1
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-RESTful==0.3.4
Flask-Script==0.3.2
Flask-Security==1.6.9
Flask-Social==1.6.2
Flask-SQLAlchemy==0.15
Flask-WTF==0.9.3
FormEncode==1.3.0
funcsigs==0.4
functools32==3.2.3.post2
futures==3.0.5
GeoAlchemy2==0.2.6
geopy==1.11.0
gevent==1.0.2
github3.py==0.7.1
gnureadline==6.3.3
google-api-python-client==1.1
googlemaps==2.4.2
greenlet==0.4.9
gunicorn==19.6.0
helper==2.4.1
httplib2==0.8
idna==2.0
ipaddress==1.0.16
ipykernel==4.1.0
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.23
Jinja2==2.8
jmespath==0.7.1
jsmin==2.0.2
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
kombu==3.0.29
line-profiler==1.0
lockfile==0.9.1
Mako==1.0.1
mandrill==1.0.57
MarkupSafe==0.23
marshmallow==2.6.1
marshmallow-sqlalchemy==0.8.1
meld3==1.0.2
mistune==0.7.1
mock==0.8.0
nbconvert==4.0.0
nbformat==4.0.1
ndg-httpsclient==0.4.0
newrelic==2.60.0.46
newrelic-plugin-agent==1.3.0
nose==1.3.7
nose-timer==0.5.0
notebook==4.0.6
oauth2==1.5.211
oauth2client==1.1
oauthlib==0.6.0
paramiko==1.16.0
parse-rest==0.2.20141004
passlib==1.6.1
PasteDeploy==1.5.2
-e git+https://github.com/TrustMile/flask-sqlalchemy.git#egg=FlaskSQLAlchemy
lxml==3.6.0
path.py==8.1.2
pbr==1.6.0
pexpect==4.0.1
pickleshare==0.5
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5
pusher==1.2.3
pushwoosh==1.0.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
Pygments==2.0.2
pypushwoosh==0.2.0
pysftp==0.2.8
python-dateutil==2.4.2
python-editor==0.3
python-gflags==2.0
python-twitter==1.1
pytz==2015.4
PyYAML==3.11
pyzmq==14.7.0
redis==2.10.5
redlock==1.2.0
regex==2016.1.10
repoze.lru==0.6
requests==2.6.0
requests-oauthlib==0.4.0
rsa==3.1.4
Shapely==1.5.13
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
smtpapi==0.2.0
SQLAlchemy==1.0.13
swagger-py-codegen==0.1.15
termcolor==1.1.0
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
translationstring==1.3
uritemplate.py==0.2.0
urllib3==1.6
-e git+https://github.com/unbit/uwsgi.git@b608eb1772641d525bfde268fe9d6d8d0d5efde7#egg=uWSGI
validate-email==1.3
venusian==1.0
waitress==0.8.9
webassets==0.7
WebHelpers==1.3
WebOb==1.5.1
Werkzeug==0.10.4
WTForms==1.0.5
yappi==0.94
zope.deprecation==4.1.2
zope.interface==4.1.2
